requirements.txt
# --------- pytorch --------- #
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4
pytorch-metric-learning>=2.0.0
# faiss-gpu~=1.7.2
faiss-cpu~=1.7.2

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
wandb>=0.16.0
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- visualizations --------- #

matplotlib~=3.7.0
seaborn~=0.12.0

# --------- data --------- #

pandas~=2.0.3
polars==0.19.12
numpy~=1.25.1
omegaconf~=2.3.0
scipy~=1.11.1
pyarrow~=12.0.1
scikit-learn~=1.3.0

# --------- others --------- #
pyrootutils~=1.0.4
pre-commit      # hooks for applying linters on commit
rich~=13.4.2
pytest~=7.4.0
tqdm~=4.65.0
# sh            # for running bash commands in some tests (linux/macos only)

packaging~=23.1
setuptools~=58.1.0
motion-learning-toolbox>=1.0.6

PyYAML~=6.0
Jinja2~=3.1.2
python-dotenv~=1.0.0